+++ snowflake-distributed-service-impl/src/main/java/com/dwarfeng/sfds/impl/service/telqos/GenCommand.java
@@ -8,32 +8,26 @@ import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.Option;
 import org.springframework.stereotype.Component;
 
-import java.io.BufferedWriter;
-import java.io.File;
-import java.io.FileWriter;
 import java.util.ArrayList;
 import java.util.List;
 
 @Component
-public class GenCsvCommand extends CliCommand {
+public class GenCommand extends CliCommand {
 
-    private static final String IDENTITY = "gencsv";
+    private static final String IDENTITY = "gen";
     private static final String DESCRIPTION = "生成ID并导出到CSV操作";
 
     private static final String CMD_OPTION_NUMBER = "n";
     private static final String CMD_OPTION_FILE = "f";
 
-    private static final String CMD_LINE_SYNTAX_INTERACTIVELY = "gencsv";
-    private static final String CMD_LINE_SYNTAX_NUMBER = "gencsv -" + CMD_OPTION_NUMBER + " [number]";
-    private static final String CMD_LINE_SYNTAX_FILE = "gencsv -" + CMD_OPTION_FILE + " [file-path]";
+    private static final String CMD_LINE_SYNTAX_INTERACTIVELY = "gen";
+    private static final String CMD_LINE_SYNTAX_NUMBER = "gen -" + CMD_OPTION_NUMBER + " [number]";
     private static final String CMD_LINE_SYNTAX = CMD_LINE_SYNTAX_INTERACTIVELY + System.lineSeparator() +
-            CMD_LINE_SYNTAX_NUMBER + System.lineSeparator() + CMD_LINE_SYNTAX_FILE;
-
-    private static final String DEFAULT_FILE_NAME = "export.csv";
+            CMD_LINE_SYNTAX_NUMBER;
 
     private final LongIdService longIdService;
 
-    public GenCsvCommand(LongIdService longIdService) {
+    public GenCommand(LongIdService longIdService) {
         super(IDENTITY, DESCRIPTION, CMD_LINE_SYNTAX);
         this.longIdService = longIdService;
     }
@@ -43,8 +37,6 @@ public class GenCsvCommand extends CliCommand {
         List<Option> list = new ArrayList<>();
         list.add(Option.builder(CMD_OPTION_NUMBER).optionalArg(true).hasArg(true).type(Number.class)
                 .argName("number").desc("生成数量").build());
-        list.add(Option.builder(CMD_OPTION_FILE).optionalArg(true).hasArg(true).type(String.class)
-                .argName("file-path").desc("生成的CSV的路径").build());
         return list;
     }
 
@@ -52,36 +44,27 @@ public class GenCsvCommand extends CliCommand {
     protected void executeWithCmd(Context context, CommandLine cmd) throws TelqosException {
         try {
             int number;
-            File file;
             if (cmd.hasOption(CMD_OPTION_NUMBER)) {
                 number = ((Number) cmd.getParsedOptionValue(CMD_OPTION_NUMBER)).intValue();
             } else {
                 context.sendMessage("请输入需要生成的 ID 的数量:");
                 number = Integer.parseInt(context.receiveMessage());
             }
-            if (cmd.hasOption(CMD_OPTION_FILE)) {
-                file = new File(((String) cmd.getParsedOptionValue(CMD_OPTION_FILE)));
-            } else {
-                context.sendMessage("请输入需要导出的 CSV 的位置:");
-                file = new File(context.receiveMessage());
+            int digits = digits(number - 1);
+            List<Long> longs = longIdService.nextLongId(number);
+            for (int i = 0; i < longs.size(); i++) {
+                context.sendMessage(String.format("%-" + digits + "d: %d", i, longs.get(i)));
             }
-            if (file.isDirectory()) {
-                file = new File(file, DEFAULT_FILE_NAME);
-            }
-            try (
-                    FileWriter fw = new FileWriter(file);
-                    BufferedWriter bw = new BufferedWriter(fw, 4096)
-            ) {
-                List<Long> longs = longIdService.nextLongId(number);
-                for (Long aLong : longs) {
-                    bw.write(Long.toString(aLong));
-                    bw.write(System.lineSeparator());
-                }
-                bw.flush();
-            }
-            context.sendMessage("CSV文件已导出至 " + file.getAbsolutePath());
         } catch (Exception e) {
             throw new TelqosException(e);
         }
+    }
+
+    private int digits(int target) {
+        int digits = 0;
+        do {
+            digits++;
+        } while ((target /= 10) > 0);
+        return digits;
     }
 }